Fix: develop 오류 수정 (merger)

+++ backend/lighthouse/src/main/java/com/ssafy/lighthouse/domain/common/repository/SidoRepository.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface SidoRepository extends JpaRepository<Sido, Integer> {
+public interface SidoRepository extends JpaRepository<Sido, Long> {
     @Query("SELECT new com.ssafy.lighthouse.domain.common.dto.SidoDto(s.id, s.name) FROM Sido s")
     List<SidoDto> findAllSidoList();
 }
